+++ lib/src/runner/browser/browser_manager.dart
@@ -11,7 +11,7 @@ import 'package:stream_channel/stream_channel.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
 
 import '../../backend/runtime.dart';
-import '../../backend/suite_platform.dart';
+import '../../util/io.dart';
 import '../../util/stack_trace_mapper.dart';
 import '../application_exception.dart';
 import '../configuration/suite.dart';
@@ -242,7 +242,7 @@ class BrowserManager {
       });
 
       try {
-        controller = deserializeSuite(path, new SuitePlatform.current(_runtime),
+        controller = deserializeSuite(path, currentPlatform(_runtime),
             suiteConfig, await _environment, suiteChannel, message);
 
         controller.channel("test.browser.mapper").sink.add(mapper?.serialize());